pyproject.toml
[project]
name = "bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gpib-ctypes>=0.3.0",
    "psutil>=7.2.2",
    "pymeasure>=0.15.0",
    "pyside6>=6.10.2",
    "pyvisa-py>=0.8.1",
    "zeroconf>=0.148.0",
]
